requirements.txt
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1
Faker~=30.8.0
alembic~=1.13.3